# app/docs/swagger.yaml
definitions:
...
  domain.Depend:
    properties:
      name:
        type: string
      version:
        type: string
    type: object
...
  domain.FbAuthentication:
    properties:
      authSupplier:
        description: 验证供应商:openid、github、google
        type: string
      config:
        description: 身份验证配置对应的配置项：供应商id、appID、appSecret、服务发现地址、重定向url等
        type: string
      id:
        type: integer
      name:
        description: 身份验证名称
        type: string
      switchState:
        description: '开关状态: 0-全部关闭 1-cookie 2-token 3-全部cookie、token都开启'
        type: string
    type: object
...
  domain.HookStruct:
    properties:
      depend:
        items:
          $ref: '#/definitions/domain.Depend'
        type: array
      input: {}
      path:
        description: 钩子的路径 钩子分四种operation、auth、customize、global 以这4个为一级目录，二级目录为钩子的类型(operation的则以operation的名称为二级目录，钩子类型为三级目录)
        type: string
      script:
        type: string
      scriptType:
        type: string
      switch:
        type: boolean
      type:
        type: string
    type: object
...
host: localhost:9123
info:
  contact: {}
  description: FireBoom-server
  title: FireBoom-server Swagger API
  version: "1.0"
paths:
  /api/v1/auth:
    get:
      consumes:
      - application/json
      description: 查询身份验证信息
      responses:
        "200":
          description: 查询成功
          schema:
            items:
              $ref: '#/definitions/domain.FbAuthentication'
            type: array
        "400":
          description: 查询失败
      tags:
      - auth
    post:
      consumes:
      - application/json
      description: 添加身份验证信息
      parameters:
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      - description: 验证供应商类型
        in: formData
        name: authSupplier
        required: true
        type: integer
      - description: 开关
        in: formData
        items:
          type: string
        name: switchState
        required: true
        type: array
      - description: 验证配置
        in: formData
        name: config
        required: true
        type: string
      responses:
        "200":
          description: 添加成功
        "400":
          description: 添加失败
      tags:
      - auth
    put:
      consumes:
      - application/json
      description: 修改身份验证信息
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: integer
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      - description: 验证供应商类型
        in: formData
        name: authSupplier
        required: true
        type: integer
      - description: 开关
        in: formData
        name: switchState
        required: true
        type: string
      - description: 验证配置
        in: formData
        name: config
        required: true
        type: string
      responses:
        "200":
          description: 修改成功
        "400":
          description: 修改失败
      tags:
      - auth
  /api/v1/auth/:id:
    delete:
      consumes:
      - application/json
      description: 删除身份验证信息
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      tags:
      - auth
...
  /api/v1/auth/getDiscoveryURL:
    get:
      consumes:
      - application/json
      description: 获取服务发现地址
      parameters:
      - description: issuer地址
        in: formData
        name: issuer
        required: true
        type: string
      - description: 供应商id
        in: formData
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              type: string
            type: array
        "400":
          description: 获取失败
      tags:
      - auth
...
  /api/v1/auth/hooks:
    get:
      consumes:
      - application/json
      description: 获取钩子内容
      parameters:
      - description: 钩子路径
        in: formData
        name: path
        required: true
        type: string
      responses:
        "200":
          description: 查询成功
          schema:
            items:
              $ref: '#/definitions/domain.HookStruct'
            type: array
        "400":
          description: 查询失败
      tags:
      - hook
    post:
      consumes:
      - application/json
      description: 执行钩子
      parameters:
      - description: 路径
        in: formData
        name: path
        required: true
        type: string
      - description: 依赖数组,swagger的问题 把depend1中的1给去掉哈
        in: formData
        items:
          type: object
        name: depend1
        required: true
        type: array
      - description: 输入
        in: formData
        name: input
        required: true
        type: object
      - description: 钩子内容
        in: formData
        name: script
        required: true
        type: string
      - description: 钩子类型
        in: formData
        name: scriptType
        required: true
        type: string
      responses:
        "200":
          description: 修改成功
        "400":
          description: 修改失败
      tags:
      - hook
...
  /api/v1/auth/redirectUrl:
    get:
      consumes:
      - application/json
      description: 获取重定向url
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              type: string
            type: array
        "400":
          description: 获取失败
      tags:
      - auth
    post:
      consumes:
      - application/json
      description: 修改重定向url
      parameters:
      - description: 重定向url数组
        in: formData
        items:
          type: string
        name: redirectURLs
        required: true
        type: array
      responses:
        "200":
          description: 修改成功
        "400":
          description: 获取失败
      tags:
      - auth
...
produces:
- application/json
swagger: "2.0"